feat(plugin-hono-server): export the current-user endpoint registrar so a bare-adapter host can supply them (cloud#924)
`/api/v1/auth/me/permissions`, `/auth/me/localization` and `/me/apps` are the
platform's SOLE supply — `packages/rest` and `packages/runtime` register no
`/me/*` route, the objectui console reads the first for its whole permission
layer and the second for regional defaults, and `core/security/auth-gate.ts`
allow-lists the last two as endpoints a gated user MUST still reach.

#4073/#4079 freed them from `registerStandardEndpoints` but left the supply
welded to `HonoServerPlugin`. A host that stands up a bare `HonoHttpServer` and
registers it as `http.server` itself — cloud's default Vercel/serverless
`bootKernel` branch, whose `OS_NODE_SERVE=1` sibling mounts the real plugin —
got no provider at all, so the console's FLS/`apiOperations` had a server-side
answer on one startup path and a 404 on the other (cloud#924).

Registration is a function of a Hono app plus a service locator, not of owning
the listening socket, so it moves to `./current-user-endpoints` and is exported:

- `registerCurrentUserEndpoints({ rawApp, ctx, prefix? })` where `ctx` is any
  `{ getService, logger }` — a `PluginContext` satisfies it structurally.
- Idempotent: returns `false` and registers nothing when all three paths are
  already served, so a host may pre-register on the raw app AND mount the
  plugin without the plugin shadowing it with dead duplicates. `every`, not
  `some`, so a host owning just one path still gets the other two.
- `makeExecutionContextResolver` and the four permission-map shaping helpers
  move with it (same package-root exports, no renames) — they exist only to
  shape this endpoint's `objects` map.

The plugin now delegates from the same `kernel:ready` hook, in the same order
relative to the CRUD/discovery block, so `os serve` and every plugin-mounting
host are byte-identical.

Tests: 137 passed in plugin-hono-server (5 new — bare-app mount + answer,
non-default prefix, idempotence, partial-host re-registration, and a host
pre-register + plugin mount that must yield ONE registration owned by the host).
client 200, http-conformance 46, runtime hono/route-parity/ready 16 — all
unchanged. tsc + eslint clean; `check:authz-resolver` still passes.
