+++ packages/plugins/plugin-hono-server/src/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2025 ObjectStack. Licensed under the Apache-2.0 license.
 
 export * from './hono-plugin';
+export * from './current-user-endpoints';
 export * from './adapter';
 export * from './pattern-matcher';
 export * from './route-pattern';
